eventservice/requirements.txt
kink==0.8.0
pika==1.3.2
psycopg==3.1.19
psycopg2-binary==2.9.1
psycopg-pool==3.2.2
SQLAlchemy==2.0.31
python-dotenv==1.0.1
